move lib functions to utils
